+++ Connect Four/Classes/Players/Player.cs
@@ -6,9 +6,9 @@ namespace Connect_Four.Classes.Players;
 public abstract class Player : IPlayer
 {
     public string Name { get; set; }
-    public short PlayerNumber { get; set; }
+    public byte PlayerNumber { get; set; }
 
-    protected Player(string name, short playerNumber)
+    protected Player(string name, byte playerNumber)
     {
         Name = name;
         PlayerNumber = playerNumber;
